feat: use `InteractionBase` to expand support for all interactions even buttons for `Nexus.R` (and also `autoDefer`).
