+++ src/main/java/pw/mihou/nexus/features/commons/Deferrable.kt
@@ -2,12 +2,14 @@ package pw.mihou.nexus.features.commons
 
 import org.javacord.api.entity.message.MessageFlag
 import org.javacord.api.interaction.Interaction
+import org.javacord.api.interaction.InteractionBase
 import org.javacord.api.interaction.callback.InteractionOriginalResponseUpdater
 import org.javacord.api.util.logging.ExceptionLogger
 import pw.mihou.nexus.Nexus
 import pw.mihou.nexus.configuration.modules.Cancellable
 import pw.mihou.nexus.features.command.responses.NexusAutoResponse
 import pw.mihou.nexus.features.messages.NexusMessage
+import pw.mihou.nexus.features.react.React
 import java.time.Instant
 import java.util.concurrent.CompletableFuture
 import java.util.concurrent.atomic.AtomicBoolean
@@ -15,8 +17,8 @@ import java.util.concurrent.atomic.AtomicReference
 import java.util.function.Function
 
 object Deferrable {
-    internal fun autoDefer(
-        event: NexusInteractionEvent<*, *>,
+    internal fun <Interaction: InteractionBase> autoDefer(
+        interaction: Interaction,
         updater: AtomicReference<CompletableFuture<InteractionOriginalResponseUpdater>?>,
         ephemeral: Boolean,
         response: Function<Void?, NexusMessage>
@@ -24,13 +26,13 @@ object Deferrable {
         var task: Cancellable? = null
         val deferredTaskRan = AtomicBoolean(false)
         if (updater.get() == null) {
-            val timeUntil = Instant.now().toEpochMilli() - event.interaction.creationTimestamp
+            val timeUntil = Instant.now().toEpochMilli() - interaction.creationTimestamp
                 .minusMillis(Nexus.configuration.global.autoDeferAfterMilliseconds)
                 .toEpochMilli()
 
             task = Nexus.launch.scheduler.launch(timeUntil) {
                 if (updater.get() == null) {
-                    event.respondLaterEphemerallyIf(ephemeral).exceptionally(ExceptionLogger.get())
+                    updater.updateAndGet { interaction.respondLater(ephemeral) }!!.exceptionally(ExceptionLogger.get())
                 }
                 deferredTaskRan.set(true)
             }
@@ -44,7 +46,7 @@ object Deferrable {
                 }
                 @Suppress("NAME_SHADOWING") val updater = updater.get()
                 if (updater == null) {
-                    val responder = event.respondNow()
+                    val responder = interaction.createImmediateResponder()
                     if (ephemeral) {
                         responder.setFlags(MessageFlag.EPHEMERAL)
                     }
@@ -68,5 +70,44 @@ object Deferrable {
             }
         }
         return future
+    }
+}
+
+/**
+ * Automatically answers either deferred or non-deferred based on circumstances, to configure the time that it should
+ * consider before deferring (this is based on time now - (interaction creation time - auto defer time)), you can
+ * modify [pw.mihou.nexus.configuration.modules.NexusGlobalConfiguration.autoDeferAfterMilliseconds].
+ *
+ * For slash commands and context menus, we recommend using the Nexus event's methods instead which will enable better coordination
+ * with middlewares.
+ *
+ * @param ephemeral whether to respond ephemerally or not.
+ * @param response the response to send to Discord.
+ * @return the response from Discord.
+ */
+fun <Interaction: InteractionBase> Interaction.autoDefer(ephemeral: Boolean, response: Function<Void?, NexusMessage>): CompletableFuture<NexusAutoResponse> =
+    Deferrable.autoDefer(this, AtomicReference(null), ephemeral, response)
+
+/**
+ * An experimental feature to use the new Nexus.R rendering mechanism to render Discord messages
+ * with a syntax similar to a template engine that sports states (writables) that can easily update message
+ * upon state changes.
+ *
+ * This internally uses [autoDefer] to handle sending of the response, ensuring that we can handle long-running renders
+ * and others that may happen due to situations such as data fetching, etc.
+ *
+ * @param ephemeral whether to send the response as ephemeral or not.
+ * @param react the entire procedure over how rendering the response works.
+ */
+@JvmSynthetic
+fun <Interaction: InteractionBase> Interaction.R(ephemeral: Boolean, react: React.() -> Unit): CompletableFuture<NexusAutoResponse> {
+    val r = React(this.api, React.RenderMode.Interaction)
+    return autoDefer(ephemeral) {
+        react(r)
+
+        return@autoDefer r.message!!
+    }.thenApply {
+        r.resultingMessage = it.getOrRequestMessage().join()
+        return@thenApply it
     }
 }
